fix: address code-review findings for clipboard copy feature (#107)
M1: move CopyToClipboard call inside tea.Cmd closure in copySelectedText so
clipboard I/O is async and does not block the Bubble Tea Update goroutine.
S1: rename CopyToClipboard -> copyToClipboard (unexported; only internal callers).
S2: move copySelectedText from keys.go to clipboard.go where it belongs;
removes the fmt import that was added to keys.go solely for this helper.
S3: full-module go test ./... confirms x/sys v0.43.0 introduces no regressions.
C2: document no-parallel-mutation constraint on copyFn/isTTYFn/stderrWriter
vars directly in clipboard.go, close to the code it governs.
C3: add best-effort comment on OSC 52 delivery in alt-screen environments.
Tests: add C2-04 to verify copyFn is not called before cmd() is invoked (M1).
